+++ src/main/java/com/arangodb/entity/arangosearch/analyzer/StemAnalyzer.java
@@ -26,7 +26,10 @@ import com.arangodb.entity.arangosearch.AnalyzerType;
 import java.util.Objects;
 
 /**
+ * An Analyzer capable of stemming the text, treated as a single token, for supported languages.
+ *
  * @author Michele Rastelli
+ * @see <a href= "https://www.arangodb.com/docs/stable/arangosearch-analyzers.html#stem">API Documentation</a>
  */
 public class StemAnalyzer extends SearchAnalyzer {
     public StemAnalyzer() {