+++ ChangeLog.md
@@ -6,6 +6,7 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/) a
 
 ## [Unreleased]
 
+- bugfix VelocyJack deserialization
 - bugfix `allowImplicit` parameter in stream transactions
 - added `peakMemoryUsage` to aql statistics