+++ ChangeLog.md
@@ -6,6 +6,8 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/) a
 
 ## [Unreleased]
 
+## [6.6.2] - 2020-04-07
+
 - bugfix VelocyJack deserialization
 - bugfix `allowImplicit` parameter in stream transactions
 - added `peakMemoryUsage` to aql statistics